Move public classes under package 'com.adobe.marketing.mobile.edge.media'
