+++ code/edgemedia/src/phone/java/com/adobe/marketing/mobile/edge/media/MediaTracker.java
@@ -9,7 +9,7 @@
   governing permissions and limitations under the License.
 */
 
-package com.adobe.marketing.mobile;
+package com.adobe.marketing.mobile.edge.media;
 
 import java.util.Map;
 import org.jetbrains.annotations.NotNull;
