+++ code/edgemedia/src/phone/java/com/adobe/marketing/mobile/edge/media/MediaTrackerEventGenerator.java
@@ -9,8 +9,11 @@
   governing permissions and limitations under the License.
 */
 
-package com.adobe.marketing.mobile;
+package com.adobe.marketing.mobile.edge.media;
 
+import com.adobe.marketing.mobile.AdobeCallback;
+import com.adobe.marketing.mobile.Event;
+import com.adobe.marketing.mobile.EventType;
 import com.adobe.marketing.mobile.edge.media.internal.MediaObject;
 import com.adobe.marketing.mobile.services.Log;
 import java.util.Calendar;
