+++ code/edgemedia/src/test/java/com/adobe/marketing/mobile/edge/media/TestMediaTrackerEventGenerator.java
@@ -9,8 +9,10 @@
   governing permissions and limitations under the License.
 */
 
-package com.adobe.marketing.mobile;
+package com.adobe.marketing.mobile.edge.media;
 
+import com.adobe.marketing.mobile.AdobeCallback;
+import com.adobe.marketing.mobile.Event;
 import java.util.Calendar;
 import java.util.Map;
